+++ solutions/ColumnFormatter/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Start Flow Wizard
 - StandardColorsDropdown (for use in wizards)
 - IconsDropdown (for use in wizards)
+- Line Numbers in the editor (toggle in the property pane)
 
 ### Changed
 - Updated to [SharePoint Framework](https://github.com/SharePoint/sp-dev-docs/wiki/Release-Notes-for-SPFx-Package-Version-1.4.1) from 1.4.0 to 1.4.1